Got things working better
I believe the root of the problem was the BufferedReader. I changed it
so BufferedReader is never used. If you want to read stuff, just use the
DataInputStream, convert things to char (if you're reading text), and
store it in a variable. "/r/n" will indicate and end of line.